refactor: :recycle: Fix pipes not properly getting closed (2/2)
