+++ include/minishell_type.h
@@ -1,6 +1,9 @@
 #ifndef MINISHELL_TYPE_H
 # define MINISHELL_TYPE_H
 
+# include <stdbool.h>
+# include <sys/types.h>
+
 typedef enum e_procflag
 {
 	PIPE_READ = 0,
@@ -10,5 +13,7 @@ typedef enum e_procflag
 	OUTPUT = 1,
 	IO_SIZE = 2,
 }	t_procflag;
+
+typedef int	t_fd;
 
 #endif
